pyproject.toml
[tool.poetry]
name = "robotics"
version = "0.1.0"
description = ""
authors = ["Federico Barcelona <unizar@federico.barcelona>"]

[tool.poetry.dependencies]
python = "^3.5"
numpy = "~1.18.5"
matplotlib = "~3.0.3"
spidev = "^3.3"
opencv-python = "4.2.0.34"

[tool.poetry.dev-dependencies]
mamba = "^0.11.2"
PyHamcrest = "^2.0.3"
flake8 = "~3.9.2"
doublex = "^1.9.3"

[tool.poetry.scripts]
robot = 'robotics.__main__:run'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
